+++ src/Functions.Rpc.Server/IWorkerProcessChannel.cs
@@ -1,14 +1,15 @@
 // Copyright (c) .NET Foundation. All rights reserved.
 // Licensed under the MIT License. See License.txt in the project root for license information.
 
+using System.Threading;
 using System.Threading.Tasks;
 
 namespace Microsoft.Azure.WebJobs.Script.Workers
 {
-    internal interface IWorkerChannel
+    internal interface IWorkerProcessChannel : IWorkerChannel
     {
-        string Id { get; }
+        IWorkerProcess WorkerProcess { get; }
 
-        Task<WorkerStatus> GetWorkerStatusAsync();
+        Task StartWorkerProcessAsync(CancellationToken cancellationToken = default);
     }
 }